Texturing for greedy meshing

+++ BlocksEngine/include/BlocksEngine/Vertex.h
@@ -23,6 +23,6 @@ namespace BlocksEngine
             float v;
         } tex;
 
-        unsigned short texIndex;
+        unsigned int texIndex;
     };
 }
